program image scaffolding
