+++ server/app/views/admin/programs/ProgramImageFragment.html
@@ -1,0 +1,5 @@
+<!--/*@thymesVar id="templateGlobals" type="views.shared.TemplateGlobals"*/-->
+<!--/*@thymesVar id="model" type="views.admin.programs.ProgramImagePageViewModel"*/-->
+<th:block th:fragment="content">
+  <div class="padding-x-4 padding-y-3"></div>
+</th:block>
