+++ server/app/views/admin/programs/ProgramImagePageViewModel.java
@@ -1,0 +1,6 @@
+package views.admin.programs;
+
+import views.BaseViewModel;
+
+/** View model for the Thymeleaf program image admin page (scaffold). */
+public record ProgramImagePageViewModel() implements BaseViewModel {}
